Allow CMUpdater to select updates
CMUpdater has new/all option and it filters updates by modification date.

+++ index.php
@@ -57,19 +57,15 @@
            $device = $postJson['params']['device'];
            $devicePath = realpath('./_builds/'.$device);
            if (file_exists($devicePath)) {
-               $after = 0;
                if (array_key_exists('source_incremental', $postJson['params'])) {
                    $source_incremental = $postJson['params']['source_incremental'];
                    if (!empty($source_incremental)) {
                        $mc = Flight::mc();
                        $source_zip = $mc->get($source_incremental);
-                       if ($source_zip) {
-                           if (file_exists($source_zip)) {
-                               $after = filemtime($source_zip);
-                           } else {
-                               $mc->delete($source_zip);
-                               $mc->delete($source_incremental);
-                           }
+                       if ($source_zip && !file_exists($source_zip)) {
+                           $mc->delete($source_zip);
+                           $mc->delete($source_incremental);
+                           $source_zip = NULL;
                        }
                    }
                }
@@ -77,7 +73,7 @@
                if (array_key_exists('channels', $postJson['params'])) {
                    $channels = $postJson['params']['channels'];
                }
-               $tokens = new TokenCollection($channels, $devicePath, $req->base, $device, $after);
+               $tokens = new TokenCollection($channels, $devicePath, $req->base, $device);
                $ret['result'] = $tokens->getUpdateList();
            }
         }
